requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.11.7
fastapi-cache2==0.2.2
redis==4.6.0
SQLAlchemy==1.4.48
alembic==1.11.1
python-dotenv==1.0.0
pypdf==4.2.0
requests==2.32.3